refactor: group endpoint wiring by feature (#33)
